django_task/day2/pyproject.toml
[project]
name = "todo"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.14"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "ipython (>=9.11.0,<10.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "django-summernote (>=0.8.20.0,<0.9.0.0)",
    "django-cleanup (>=9.0.0,<10.0.0)",
    "pillow (>=12.1.1,<13.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
